+++ backend/prompt_studio/prompt_studio_index_manager/models.py
@@ -9,8 +9,8 @@ from django.dispatch import receiver
 from prompt_studio.prompt_profile_manager.models import ProfileManager
 from prompt_studio.prompt_studio_core.prompt_ide_base_tool import PromptIdeBaseTool
 from prompt_studio.prompt_studio_document_manager.models import DocumentManager
-from unstract.sdk.core.constants import LogLevel
-from unstract.sdk.core.vector_db import VectorDB
+from unstract.sdk.constants import LogLevel
+from unstract.sdk.vector_db import VectorDB
 from utils.models.base_model import BaseModel
 
 logger = logging.getLogger(__name__)
