+++ backend/tool_instance/tool_instance_helper.py
@@ -16,7 +16,7 @@ from tool_instance.exceptions import ToolSettingValidationError
 from tool_instance.models import ToolInstance
 from tool_instance.tool_processor import ToolProcessor
 from unstract.sdk.adapters.enums import AdapterTypes
-from unstract.sdk.core.tool.validator import DefaultsGeneratingValidator
+from unstract.sdk.tool.validator import DefaultsGeneratingValidator
 from unstract.tool_registry.constants import AdapterPropertyKey
 from unstract.tool_registry.dto import Spec, Tool
 from unstract.tool_registry.tool_utils import ToolUtils
